+++ src/include/db/mariadb.h
@@ -17,7 +17,7 @@ typedef struct {
 
 
 int MARIADB_connect( MARIADB_CONNECTION* db_connection, const char* host, const int port, const char* dbname, const char* user, const char* password, const char* connection_string );
-int MARIADB_exec( MARIADB_CONNECTION* db_connection, const char* sql, unsigned long sql_length, DB_QUERY* dst_result, const char* const *param_values, const int params_count, const int *param_lengths, const int *param_formats, const char* const *param_types );
+int MARIADB_exec( MARIADB_CONNECTION* db_connection, const char* sql, size_t sql_length, DB_QUERY* dst_result, const char* const *param_values, const int params_count, const int *param_lengths, const int *param_formats, const char* const *param_types );
 void MARIADB_disconnect( MARIADB_CONNECTION* db_connection );
 
 #endif
