+++ src/include/db/oracle.h
@@ -16,7 +16,7 @@ typedef struct {
 
 
 int ORACLE_connect( ORACLE_CONNECTION* db_connection, const char* host, const int port, const char* dbname, const char* user, const char* password, const char* connection_string );
-int ORACLE_exec( ORACLE_CONNECTION* db_connection, const char* sql, unsigned long sql_length, DB_QUERY* dst_result, const char* const *param_values, const int params_count, const int *param_lengths, const int *param_formats );
+int ORACLE_exec( ORACLE_CONNECTION* db_connection, const char* sql, size_t sql_length, DB_QUERY* dst_result, const char* const *param_values, const int params_count, const int *param_lengths, const int *param_formats );
 void ORACLE_disconnect( ORACLE_CONNECTION* db_connection );
 
 #endif
