+++ src/include/db/postgresql.h
@@ -17,7 +17,7 @@ typedef struct {
 
 
 int PG_connect( PG_CONNECTION* db_connection, const char* host, const int port, const char* dbname, const char* user, const char* password, const char* connection_string );
-int PG_exec( PG_CONNECTION* db_connection, const char* sql, unsigned long sql_length, DB_QUERY* dst_result, const char* const *param_values, const int params_count, const int *param_lengths, const int *param_formats, Oid *param_types );
+int PG_exec( PG_CONNECTION* db_connection, const char* sql, size_t sql_length, DB_QUERY* dst_result, const char* const *param_values, const int params_count, const int *param_lengths, const int *param_formats, Oid *param_types );
 void PG_disconnect( PG_CONNECTION* db_connection );
 
 #endif
